moved queries to stations, printing a list of all the stations
